refactor(conftest): migrate authentication fixtures to bookstore/conftest.py
- Moved authentication-related fixtures from conftest.py to bookstore/conftest.py for better organization.
- Updated the auth_file fixture to save the authentication state in a dedicated bookstore directory.
- Ensured compatibility with existing test structures while improving code clarity and maintainability.
