+++ conftest.py
@@ -3,17 +3,11 @@ import json
 import os
 import re
 from pathlib import Path
-from typing import Iterator
 
 import pytest
-from filelock import FileLock
 
-from playwright.sync_api import Page, Route, Browser
+from playwright.sync_api import Page, Route
 
-from bookstore.pages.login_page import LoginPage
-from bookstore.pages.profile_page import ProfilePage
-
-AUTH_FILE = Path(".auth/storage_state.json")
 AD_DOMAINS = [
     # Google Ad Services (the main culprits)
     "googleads.g.doubleclick.net",
@@ -107,61 +101,3 @@ def test_users() -> dict:
         "Test user credentials were not provided. Set the TEST_USERS_JSON env "
         "variable or create shared/test_data/test_users.json."
     )
-
-
-@pytest.fixture(scope="session")
-def auth_file(browser: Browser, test_users: dict, profile_name: str) -> Path:
-    """
-    Session-scoped fixture to log in once and save state.
-    Reuses existing state if available. Safe for parallel execution.
-    """
-    auth_path = (
-        AUTH_FILE
-        if profile_name == "profile1"
-        else AUTH_FILE.with_name(f"{AUTH_FILE.stem}_{profile_name}{AUTH_FILE.suffix}")
-    )
-    auth_path.parent.mkdir(exist_ok=True)
-    lock_path = auth_path.with_suffix(".lock")
-
-    with FileLock(lock_path):
-        if auth_path.is_file():
-            # Reuse existing auth state — no login needed
-            return auth_path
-
-        # Otherwise, log in and save state
-        page = browser.new_page()
-        user = test_users[profile_name]
-        login_page = LoginPage(page)
-        login_page.load()
-        login_page.login(user["email"], user["password"])
-        page.wait_for_url("**/profile", timeout=10000)
-        page.context.storage_state(path=auth_path)
-        page.close()
-
-    return auth_path
-
-
-@pytest.fixture()
-def logged_in_page(browser: Browser, auth_file: Path) -> Iterator[ProfilePage]:
-    """
-    Returns a BasePage instance from a new, pre-authenticated
-    browser context. The page is navigated to the profile page
-    to ensure it is in a valid, ready-to-use state.
-    """
-    context = browser.new_context(storage_state=auth_file)
-    page = context.new_page()
-
-    profile_page = ProfilePage(page)
-    profile_page.load()
-
-    yield profile_page
-    context.close()
-
-
-@pytest.fixture()
-def orders_cleanup(logged_in_page: ProfilePage) -> Iterator[None]:
-    """For tests that create orders, clear them before and after the run."""
-    yield
-
-    logged_in_page.load()
-    logged_in_page.delete_all_orders_if_present()
